Add more domain models, update strategies, remove unused classes

+++ src/main/java/com/utdallas/movierental/price/ChildrensPrice.java
@@ -1,18 +1,20 @@
 package com.utdallas.movierental.price;
 
+import java.math.BigDecimal;
+
 public class ChildrensPrice implements Price {
 
     private static final int EXTENDED_RENT_LENGTH = 3;
     private static final double CHARGE_AMOUNT = 1.5;
 
     @Override
-    public double getChargeAmount(int daysRented) {
+    public BigDecimal getChargeAmount(int daysRented) {
         double chargeAmount = CHARGE_AMOUNT;
         if (daysRented > EXTENDED_RENT_LENGTH) {
             chargeAmount += (daysRented - EXTENDED_RENT_LENGTH) * CHARGE_AMOUNT;
         }
 
-        return chargeAmount;
+        return BigDecimal.valueOf(chargeAmount);
     }
 
 }
