+++ src/main/java/com/utdallas/movierental/price/Price.java
@@ -1,7 +1,10 @@
 package com.utdallas.movierental.price;
 
+import java.math.BigDecimal;
+
+//TODO Separate days rented to strategy like frequent renter points?
 public interface Price {
 
-    double getChargeAmount(int daysRented);
+    BigDecimal getChargeAmount(int daysRented);
 
 }
