+++ src/main/java/com/utdallas/movierental/price/RegularPrice.java
@@ -1,17 +1,19 @@
 package com.utdallas.movierental.price;
 
+import java.math.BigDecimal;
+
 public class RegularPrice implements Price {
 
     private static final int EXTENDED_RENT_LENGTH = 2;
     private static final double CHARGE_AMOUNT = 2;
 
     @Override
-    public double getChargeAmount(int daysRented) {
+    public BigDecimal getChargeAmount(int daysRented) {
         double chargeAmount = CHARGE_AMOUNT;
         if (daysRented > EXTENDED_RENT_LENGTH) {
             chargeAmount += (daysRented - EXTENDED_RENT_LENGTH) * CHARGE_AMOUNT;
         }
 
-        return chargeAmount;
+        return BigDecimal.valueOf(chargeAmount);
     }
 }
